+++ project/src/store/current-movie/current-movie-reducer.test.ts
@@ -1,10 +1,10 @@
 import { UNKNOWN_ACTION } from '../../const/const';
 import { currentMovieInitialState } from '../../const/initial-states';
 import { testUtils } from '../../utils/mocks/test-utils';
-import { fetchCurrentMovieAction, fetchReviewsAction, fetchSimilarMoviesAction } from './current-movie-api-actions';
+import { addReviewAction, fetchCurrentMovieAction, fetchReviewsAction, fetchSimilarMoviesAction } from './current-movie-api-actions';
 import currentMovieReducer from './current-movie-reducer';
 
-const {mockCurrentMovie, mockReviews, mockSimilarMovies} = testUtils();
+const {mockCurrentMovie, mockReviews, mockReview, mockSimilarMovies} = testUtils();
 
 describe('Reducer: currentMovie', () => {
   const state = currentMovieInitialState;
@@ -32,6 +32,23 @@ describe('Reducer: currentMovie', () => {
     it('should load similar movies if fetchSimilarMovies was fulfilled', () => {
       expect(currentMovieReducer(state, {type: fetchSimilarMoviesAction.fulfilled.type, payload: mockSimilarMovies}))
         .toEqual({...state, similar: {...state.reviews, data: mockSimilarMovies, isLoaded: true}});
+    });
+  });
+
+  describe('addReviewAction test', () => {
+    it('should set isAddingReview to false if addReviewAction was fulfilled', () => {
+      expect(currentMovieReducer(state, { type: addReviewAction.fulfilled.type, payload: mockReview }))
+        .toEqual({...state, isAddingReview: false});
+    });
+
+    it('should set isAddingReview to false if addReviewAction was rejected', () => {
+      expect(currentMovieReducer(state, { type: addReviewAction.rejected.type, payload: mockReview }))
+        .toEqual({...state, isAddingReview: false});
+    });
+
+    it('should set isAddingReview to true if addReviewAction is pending', () => {
+      expect(currentMovieReducer(state, { type: addReviewAction.pending.type, payload: mockReview }))
+        .toEqual({...state, isAddingReview: true});
     });
   });
 });
