+++ project/src/store/user/user-api-actions.ts
@@ -5,10 +5,8 @@ import { dropToken, saveToken } from '../../services/token/token';
 import AppDispatch from '../../types/app-dispatch';
 import { TAuthData, TUserInfo } from '../../types/data';
 import TMovie from '../../types/movie';
-import TReview from '../../types/review';
-import { State, TReviewState } from '../../types/state';
+import { State } from '../../types/state';
 import { redirectToRoute } from '../common/common-actions';
-import { fetchCurrentMovieAction, fetchReviewsAction } from '../current-movie/current-movie-api-actions';
 
 export const fetchUserInfoAction = createAsyncThunk<TUserInfo, undefined, {
   dispatch: AppDispatch,
@@ -81,19 +79,5 @@ export const toggleFavoriteAction = createAsyncThunk<void, {id: number, status: 
   async ({id, status}, {dispatch, extra: api}) => {
     await api.post<TMovie>(`${APIRoute.Favorites}/${id}/${status}`);
     dispatch(fetchFavoritesAction());
-  },
-);
-
-export const addReviewAction = createAsyncThunk<void, TReviewState & {id: number}, {
-  dispatch: AppDispatch,
-  state: State,
-  extra: AxiosInstance
-}>(
-  ChangeAction.AddReview,
-  async ({rating, comment, id}, {dispatch, extra: api}) => {
-    await api.post<TReview[]>(`${APIRoute.Review}/${id}`, {comment, rating});
-    dispatch(fetchReviewsAction(id));
-    dispatch(fetchCurrentMovieAction(id));
-    dispatch(redirectToRoute(`${AppRoute.Movies}${id}`));
   },
 );
