+++ project/src/store/user/user-selectors.ts
@@ -8,5 +8,3 @@ export const getAuthStatus = createSelector(getUserState, (state) => state.authS
 export const getUserInfo = createSelector(getUserState, (state) => state.userInfo);
 
 export const getFavorites = createSelector(getUserState, (state) => state.favorites);
-
-export const getAddReviewState = createSelector(getUserState, (state) => state.isAddingReview);
